fix(DisputeLogics): enforce in-progress state for critical functions
This commit introduces an `inProgress` modifier to ensure that certain functions in the `DisputeLogics` and `FailLogics` contracts can only be executed when the contract is in the IN_PROGRESS state. This change enhances the contract's state management and prevents operations from being executed when they are not allowed, improving overall reliability.

Fixes #39